# src/internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.

package db

import (
	"time"

	"github.com/google/uuid"
)

type User struct {
	ID        uuid.UUID `json:"id"         db:"id"`
	Name      string    `json:"name"       db:"name"`
	Email     string    `json:"email"      db:"email"`
	CreatedAt time.Time `json:"created_at" db:"created_at"`
}

type Product struct {
	ID          uuid.UUID `json:"id"          db:"id"`
	Name        string    `json:"name"        db:"name"`
	Description *string   `json:"description" db:"description"`
	Price       float64   `json:"price"       db:"price"`
	CreatedAt   time.Time `json:"created_at"  db:"created_at"`
}

type Cart struct {
	ID        uuid.UUID `json:"id"         db:"id"`
	UserID    uuid.UUID `json:"user_id"    db:"user_id"`
	ProductID uuid.UUID `json:"product_id" db:"product_id"`
	Quantity  int32     `json:"quantity"   db:"quantity"`
	CreatedAt time.Time `json:"created_at" db:"created_at"`
}